requirements.txt
tweepy~=3.10.0
bson~=0.5.10
PyYAML~=6.0
pyspark~=3.2.1
Faker~=13.6.0
kafka~=1.3.5
pymongo~=4.1.1
typing~=3.7.4.3
elasticsearch~=8.1.3
apache-airflow-providers-cncf-kubernetes